third_party/WebKit/LayoutTests/external/wpt/webrtc/interfaces.https-expected.txt

This is a testharness.js-based test.
FAIL Main test driver promise_test: Unhandled rejection with value: object "TypeError: Cannot read property 'constructor' of undefined"
PASS Test driver for asyncInitCertificate
FAIL Test driver for asyncInitTransports assert_unreached: Failed to run asyncInitTransports: ReferenceError: RTCSctpTransport is not defined Reached unreachable code
PASS Test driver for asyncInitMediaStreamTrack
FAIL MediaStreamTrack interface: existence and properties of interface object Cannot read property 'has_extended_attribute' of undefined
PASS MediaStreamTrack interface object length
PASS MediaStreamTrack interface object name
FAIL MediaStreamTrack interface: existence and properties of interface prototype object Cannot read property 'has_extended_attribute' of undefined
PASS MediaStreamTrack interface: existence and properties of interface prototype object's "constructor" property
PASS MediaStreamTrack interface: existence and properties of interface prototype object's @@unscopables property
FAIL MediaStreamTrack interface: attribute isolated assert_true: The prototype object must have a property "isolated" expected true got false
PASS Unscopable handled correctly for isolated property on MediaStreamTrack
FAIL MediaStreamTrack interface: attribute onisolationchange assert_true: The prototype object must have a property "onisolationchange" expected true got false
PASS Unscopable handled correctly for onisolationchange property on MediaStreamTrack
PASS MediaStreamTrack must be primary interface of idlTestObjects.mediaStreamTrack
FAIL Stringification of idlTestObjects.mediaStreamTrack Cannot read property 'has_stringifier' of undefined
FAIL MediaStreamTrack interface: idlTestObjects.mediaStreamTrack must inherit property "isolated" with the proper type assert_inherits: property "isolated" not found in prototype chain
FAIL MediaStreamTrack interface: idlTestObjects.mediaStreamTrack must inherit property "onisolationchange" with the proper type assert_inherits: property "onisolationchange" not found in prototype chain
FAIL MediaStreamTrack must be primary interface of generateMediaStreamTrack('audio') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL Stringification of generateMediaStreamTrack('audio') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL MediaStreamTrack interface: generateMediaStreamTrack('audio') must inherit property "isolated" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL MediaStreamTrack interface: generateMediaStreamTrack('audio') must inherit property "onisolationchange" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL RTCPeerConnection interface: existence and properties of interface object Cannot read property 'has_extended_attribute' of undefined
PASS RTCPeerConnection interface object length
PASS RTCPeerConnection interface object name
FAIL RTCPeerConnection interface: existence and properties of interface prototype object Cannot read property 'has_extended_attribute' of undefined
PASS RTCPeerConnection interface: existence and properties of interface prototype object's "constructor" property
PASS RTCPeerConnection interface: existence and properties of interface prototype object's @@unscopables property
PASS RTCPeerConnection interface: operation createOffer(RTCOfferOptions)
PASS Unscopable handled correctly for createOffer(RTCOfferOptions) on RTCPeerConnection
PASS RTCPeerConnection interface: operation createAnswer(RTCAnswerOptions)
PASS Unscopable handled correctly for createAnswer(RTCAnswerOptions) on RTCPeerConnection
PASS RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit)
PASS Unscopable handled correctly for setLocalDescription(RTCSessionDescriptionInit) on RTCPeerConnection
PASS RTCPeerConnection interface: attribute localDescription
PASS Unscopable handled correctly for localDescription property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute currentLocalDescription assert_true: The prototype object must have a property "currentLocalDescription" expected true got false
PASS Unscopable handled correctly for currentLocalDescription property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute pendingLocalDescription assert_true: The prototype object must have a property "pendingLocalDescription" expected true got false
PASS Unscopable handled correctly for pendingLocalDescription property on RTCPeerConnection
PASS RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit)
PASS Unscopable handled correctly for setRemoteDescription(RTCSessionDescriptionInit) on RTCPeerConnection
PASS RTCPeerConnection interface: attribute remoteDescription
PASS Unscopable handled correctly for remoteDescription property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute currentRemoteDescription assert_true: The prototype object must have a property "currentRemoteDescription" expected true got false
PASS Unscopable handled correctly for currentRemoteDescription property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute pendingRemoteDescription assert_true: The prototype object must have a property "pendingRemoteDescription" expected true got false
PASS Unscopable handled correctly for pendingRemoteDescription property on RTCPeerConnection
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object])
PASS Unscopable handled correctly for addIceCandidate([object Object],[object Object]) on RTCPeerConnection
PASS RTCPeerConnection interface: attribute signalingState
PASS Unscopable handled correctly for signalingState property on RTCPeerConnection
PASS RTCPeerConnection interface: attribute iceGatheringState
PASS Unscopable handled correctly for iceGatheringState property on RTCPeerConnection
PASS RTCPeerConnection interface: attribute iceConnectionState
PASS Unscopable handled correctly for iceConnectionState property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute connectionState assert_true: The prototype object must have a property "connectionState" expected true got false
PASS Unscopable handled correctly for connectionState property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute canTrickleIceCandidates assert_true: The prototype object must have a property "canTrickleIceCandidates" expected true got false
PASS Unscopable handled correctly for canTrickleIceCandidates property on RTCPeerConnection
FAIL RTCPeerConnection interface: operation getDefaultIceServers() assert_own_property: interface object missing static operation expected property "getDefaultIceServers" missing
PASS Unscopable handled correctly for getDefaultIceServers() on RTCPeerConnection
FAIL RTCPeerConnection interface: operation getConfiguration() assert_own_property: interface prototype object missing non-static operation expected property "getConfiguration" missing
PASS Unscopable handled correctly for getConfiguration() on RTCPeerConnection
PASS RTCPeerConnection interface: operation setConfiguration(RTCConfiguration)
PASS Unscopable handled correctly for setConfiguration(RTCConfiguration) on RTCPeerConnection
PASS RTCPeerConnection interface: operation close()
PASS Unscopable handled correctly for close() on RTCPeerConnection
PASS RTCPeerConnection interface: attribute onnegotiationneeded
PASS Unscopable handled correctly for onnegotiationneeded property on RTCPeerConnection
PASS RTCPeerConnection interface: attribute onicecandidate
PASS Unscopable handled correctly for onicecandidate property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute onicecandidateerror assert_true: The prototype object must have a property "onicecandidateerror" expected true got false
PASS Unscopable handled correctly for onicecandidateerror property on RTCPeerConnection
PASS RTCPeerConnection interface: attribute onsignalingstatechange
PASS Unscopable handled correctly for onsignalingstatechange property on RTCPeerConnection
PASS RTCPeerConnection interface: attribute oniceconnectionstatechange
PASS Unscopable handled correctly for oniceconnectionstatechange property on RTCPeerConnection
PASS RTCPeerConnection interface: attribute onicegatheringstatechange
PASS Unscopable handled correctly for onicegatheringstatechange property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute onconnectionstatechange assert_true: The prototype object must have a property "onconnectionstatechange" expected true got false
PASS Unscopable handled correctly for onconnectionstatechange property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute onfingerprintfailure assert_true: The prototype object must have a property "onfingerprintfailure" expected true got false
PASS Unscopable handled correctly for onfingerprintfailure property on RTCPeerConnection
PASS RTCPeerConnection interface: operation createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions)
PASS Unscopable handled correctly for createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions) on RTCPeerConnection
PASS RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)
PASS Unscopable handled correctly for setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on RTCPeerConnection
PASS RTCPeerConnection interface: operation createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback)
PASS Unscopable handled correctly for createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback) on RTCPeerConnection
PASS RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)
PASS Unscopable handled correctly for setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on RTCPeerConnection
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback)
PASS Unscopable handled correctly for addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback) on RTCPeerConnection
PASS RTCPeerConnection interface: operation generateCertificate(AlgorithmIdentifier)
PASS Unscopable handled correctly for generateCertificate(AlgorithmIdentifier) on RTCPeerConnection
PASS RTCPeerConnection interface: operation getSenders()
PASS Unscopable handled correctly for getSenders() on RTCPeerConnection
PASS RTCPeerConnection interface: operation getReceivers()
PASS Unscopable handled correctly for getReceivers() on RTCPeerConnection
FAIL RTCPeerConnection interface: operation getTransceivers() assert_own_property: interface prototype object missing non-static operation expected property "getTransceivers" missing
PASS Unscopable handled correctly for getTransceivers() on RTCPeerConnection
PASS RTCPeerConnection interface: operation addTrack(MediaStreamTrack, MediaStream)
PASS Unscopable handled correctly for addTrack(MediaStreamTrack, MediaStream) on RTCPeerConnection
PASS RTCPeerConnection interface: operation removeTrack(RTCRtpSender)
PASS Unscopable handled correctly for removeTrack(RTCRtpSender) on RTCPeerConnection
FAIL RTCPeerConnection interface: operation addTransceiver([object Object],[object Object], RTCRtpTransceiverInit) assert_own_property: interface prototype object missing non-static operation expected property "addTransceiver" missing
PASS Unscopable handled correctly for addTransceiver([object Object],[object Object], RTCRtpTransceiverInit) on RTCPeerConnection
PASS RTCPeerConnection interface: attribute ontrack
PASS Unscopable handled correctly for ontrack property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute sctp assert_true: The prototype object must have a property "sctp" expected true got false
PASS Unscopable handled correctly for sctp property on RTCPeerConnection
PASS RTCPeerConnection interface: operation createDataChannel(USVString, RTCDataChannelInit)
PASS Unscopable handled correctly for createDataChannel(USVString, RTCDataChannelInit) on RTCPeerConnection
PASS RTCPeerConnection interface: attribute ondatachannel
PASS Unscopable handled correctly for ondatachannel property on RTCPeerConnection
PASS RTCPeerConnection interface: operation getStats(MediaStreamTrack)
PASS Unscopable handled correctly for getStats(MediaStreamTrack) on RTCPeerConnection
FAIL RTCPeerConnection interface: operation setIdentityProvider(DOMString, RTCIdentityProviderOptions) assert_own_property: interface prototype object missing non-static operation expected property "setIdentityProvider" missing
PASS Unscopable handled correctly for setIdentityProvider(DOMString, RTCIdentityProviderOptions) on RTCPeerConnection
FAIL RTCPeerConnection interface: operation getIdentityAssertion() assert_own_property: interface prototype object missing non-static operation expected property "getIdentityAssertion" missing
PASS Unscopable handled correctly for getIdentityAssertion() on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute peerIdentity assert_true: The prototype object must have a property "peerIdentity" expected true got false
PASS Unscopable handled correctly for peerIdentity property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute idpLoginUrl assert_true: The prototype object must have a property "idpLoginUrl" expected true got false
PASS Unscopable handled correctly for idpLoginUrl property on RTCPeerConnection
FAIL RTCPeerConnection interface: attribute idpErrorInfo assert_true: The prototype object must have a property "idpErrorInfo" expected true got false
PASS Unscopable handled correctly for idpErrorInfo property on RTCPeerConnection
PASS RTCPeerConnection must be primary interface of new RTCPeerConnection()
FAIL Stringification of new RTCPeerConnection() Cannot read property 'has_stringifier' of undefined
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createOffer(RTCOfferOptions)" with the proper type
PASS RTCPeerConnection interface: calling createOffer(RTCOfferOptions) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createAnswer(RTCAnswerOptions)" with the proper type
PASS RTCPeerConnection interface: calling createAnswer(RTCAnswerOptions) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setLocalDescription(RTCSessionDescriptionInit)" with the proper type
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "localDescription" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "currentLocalDescription" with the proper type assert_inherits: property "currentLocalDescription" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "pendingLocalDescription" with the proper type assert_inherits: property "pendingLocalDescription" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setRemoteDescription(RTCSessionDescriptionInit)" with the proper type
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "remoteDescription" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "currentRemoteDescription" with the proper type assert_inherits: property "currentRemoteDescription" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "pendingRemoteDescription" with the proper type assert_inherits: property "pendingRemoteDescription" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addIceCandidate([object Object],[object Object])" with the proper type
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object]) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "signalingState" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "iceGatheringState" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "iceConnectionState" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "connectionState" with the proper type assert_inherits: property "connectionState" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "canTrickleIceCandidates" with the proper type assert_inherits: property "canTrickleIceCandidates" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getDefaultIceServers()" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getConfiguration()" with the proper type assert_inherits: property "getConfiguration" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setConfiguration(RTCConfiguration)" with the proper type
PASS RTCPeerConnection interface: calling setConfiguration(RTCConfiguration) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "close()" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onnegotiationneeded" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicecandidate" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicecandidateerror" with the proper type assert_inherits: property "onicecandidateerror" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onsignalingstatechange" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "oniceconnectionstatechange" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicegatheringstatechange" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onconnectionstatechange" with the proper type assert_inherits: property "onconnectionstatechange" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onfingerprintfailure" with the proper type assert_inherits: property "onfingerprintfailure" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions)" with the proper type
PASS RTCPeerConnection interface: calling createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "generateCertificate(AlgorithmIdentifier)" with the proper type
PASS RTCPeerConnection interface: calling generateCertificate(AlgorithmIdentifier) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getSenders()" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getReceivers()" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getTransceivers()" with the proper type assert_inherits: property "getTransceivers" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addTrack(MediaStreamTrack, MediaStream)" with the proper type
PASS RTCPeerConnection interface: calling addTrack(MediaStreamTrack, MediaStream) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "removeTrack(RTCRtpSender)" with the proper type
PASS RTCPeerConnection interface: calling removeTrack(RTCRtpSender) on new RTCPeerConnection() with too few arguments must throw TypeError
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addTransceiver([object Object],[object Object], RTCRtpTransceiverInit)" with the proper type assert_inherits: property "addTransceiver" not found in prototype chain
FAIL RTCPeerConnection interface: calling addTransceiver([object Object],[object Object], RTCRtpTransceiverInit) on new RTCPeerConnection() with too few arguments must throw TypeError assert_inherits: property "addTransceiver" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "ontrack" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "sctp" with the proper type assert_inherits: property "sctp" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createDataChannel(USVString, RTCDataChannelInit)" with the proper type
PASS RTCPeerConnection interface: calling createDataChannel(USVString, RTCDataChannelInit) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "ondatachannel" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getStats(MediaStreamTrack)" with the proper type
PASS RTCPeerConnection interface: calling getStats(MediaStreamTrack) on new RTCPeerConnection() with too few arguments must throw TypeError
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setIdentityProvider(DOMString, RTCIdentityProviderOptions)" with the proper type assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: calling setIdentityProvider(DOMString, RTCIdentityProviderOptions) on new RTCPeerConnection() with too few arguments must throw TypeError assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getIdentityAssertion()" with the proper type assert_inherits: property "getIdentityAssertion" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "peerIdentity" with the proper type assert_inherits: property "peerIdentity" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "idpLoginUrl" with the proper type assert_inherits: property "idpLoginUrl" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "idpErrorInfo" with the proper type assert_inherits: property "idpErrorInfo" not found in prototype chain
PASS RTCSessionDescription interface: existence and properties of interface object
FAIL RTCSessionDescription interface object length assert_equals: wrong value for RTCSessionDescription.length expected 1 but got 0
PASS RTCSessionDescription interface object name
PASS RTCSessionDescription interface: existence and properties of interface prototype object
PASS RTCSessionDescription interface: existence and properties of interface prototype object's "constructor" property
PASS RTCSessionDescription interface: existence and properties of interface prototype object's @@unscopables property
FAIL RTCSessionDescription interface: attribute type assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
PASS Unscopable handled correctly for type property on RTCSessionDescription
FAIL RTCSessionDescription interface: attribute sdp assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
PASS Unscopable handled correctly for sdp property on RTCSessionDescription
PASS RTCSessionDescription interface: operation toJSON()
PASS Unscopable handled correctly for toJSON() on RTCSessionDescription
PASS RTCSessionDescription must be primary interface of new RTCSessionDescription({ type: 'offer' })
PASS Stringification of new RTCSessionDescription({ type: 'offer' })
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "type" with the proper type
FAIL RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "sdp" with the proper type assert_equals: expected "string" but got "object"
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "toJSON()" with the proper type
FAIL Test default toJSON operation of RTCSessionDescription assert_equals: expected "string" but got "object"
PASS RTCIceCandidate interface: existence and properties of interface object
FAIL RTCIceCandidate interface object length assert_equals: wrong value for RTCIceCandidate.length expected 0 but got 1
PASS RTCIceCandidate interface object name
PASS RTCIceCandidate interface: existence and properties of interface prototype object
PASS RTCIceCandidate interface: existence and properties of interface prototype object's "constructor" property
PASS RTCIceCandidate interface: existence and properties of interface prototype object's @@unscopables property
FAIL RTCIceCandidate interface: attribute candidate assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
PASS Unscopable handled correctly for candidate property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute sdpMid assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
PASS Unscopable handled correctly for sdpMid property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute sdpMLineIndex assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
PASS Unscopable handled correctly for sdpMLineIndex property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute foundation assert_true: The prototype object must have a property "foundation" expected true got false
PASS Unscopable handled correctly for foundation property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute component assert_true: The prototype object must have a property "component" expected true got false
PASS Unscopable handled correctly for component property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute priority assert_true: The prototype object must have a property "priority" expected true got false
PASS Unscopable handled correctly for priority property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute ip assert_true: The prototype object must have a property "ip" expected true got false
PASS Unscopable handled correctly for ip property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute protocol assert_true: The prototype object must have a property "protocol" expected true got false
PASS Unscopable handled correctly for protocol property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute port assert_true: The prototype object must have a property "port" expected true got false
PASS Unscopable handled correctly for port property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute type assert_true: The prototype object must have a property "type" expected true got false
PASS Unscopable handled correctly for type property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute tcpType assert_true: The prototype object must have a property "tcpType" expected true got false
PASS Unscopable handled correctly for tcpType property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute relatedAddress assert_true: The prototype object must have a property "relatedAddress" expected true got false
PASS Unscopable handled correctly for relatedAddress property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute relatedPort assert_true: The prototype object must have a property "relatedPort" expected true got false
PASS Unscopable handled correctly for relatedPort property on RTCIceCandidate
FAIL RTCIceCandidate interface: attribute usernameFragment assert_true: The prototype object must have a property "usernameFragment" expected true got false
PASS Unscopable handled correctly for usernameFragment property on RTCIceCandidate
PASS RTCIceCandidate interface: operation toJSON()
PASS Unscopable handled correctly for toJSON() on RTCIceCandidate
FAIL RTCIceCandidate must be primary interface of new RTCIceCandidate({ sdpMid: 1 }) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL Stringification of new RTCIceCandidate({ sdpMid: 1 }) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "candidate" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "sdpMid" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "sdpMLineIndex" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "foundation" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "component" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "priority" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "ip" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "protocol" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "port" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "type" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "tcpType" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "relatedAddress" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "relatedPort" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "usernameFragment" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "toJSON()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
Harness: the test ran to completion.
